Move store folder to src
