+++ sonar-project.properties
@@ -16,4 +16,4 @@ sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 
 sonar.test.inclusion=**/*.test.tsx,**/*.test.ts
 
-sonar.coverage.exclusions=**/*.test.tsx,src/main.tsx,**/*.test.ts,src/setUpTests.ts
+sonar.coverage.exclusions=**/*.test.tsx,src/main.tsx,**/*.test.ts,src/setupTests.ts
