+++ src/game/pages/GamesListPage/GamesListPage.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import GameListPage from "./GamesListPage";
 import { Provider } from "react-redux";
-import { store } from "../../store";
+import { store } from "../../../store";
 
 describe("Given the GameListPage component", () => {
   describe("When rendered", () => {
