+++ src/main.tsx
@@ -7,7 +7,7 @@ import "@fontsource/poppins/700.css";
 import "@fontsource/poppins/500.css";
 import "@fontsource/poppins/300.css";
 import AppRouter from "./router";
-import { store } from "./game/store";
+import { store } from "./store";
 import "./styles/styles.css";
 
 const root = document.querySelector(".root");
